Fixing queries on search: editors saw both published and draft

+++ core/src/main/java/org/fao/geonet/kernel/metadata/draft/MetadataShared.java
@@ -95,7 +95,9 @@ public class MetadataShared implements ApplicationListener<MetadataShare> {
                                     original.getId(),
                                     event.getOp().getId().getOperationId());
                     
-                    operationAllowedRepository.delete(operationAllowed);
+                    if(operationAllowed != null) {
+                        operationAllowedRepository.delete(operationAllowed);
+                    }
                     break;
                 default:
                     //There is a case for this?
